I fixed a small bug in MosaicDetection.java
MosaicClassification classifies and evaluates the results obtained in the classification. It works OK in Windows but I have to try it in Linux.
If (the annot-file path is empty) it only shows the results in the mosaic image
else, it shows the results and evaluate them. It shows the TP,FN,FP.