FN=P-TP so it's not necessary the comparison with the patches from annot-file. (I removed all this)
Here, it only studies the Neurons Classified and they are compared with the Neurons of Annot-file.

I have kept in mind three different cases:

1) one Cn (Neuron Classified) overloaps with 2 or more An (Neuron of Annot-file)
     if(area (Cn?An1)> area(Cn?An2)) Cn==> An1
     else      Cn==> An2

2) one An overlaps with 2 or more Cn
     if(area(Cn1?An)> area(Cn2?An))  An==> Cn1
     else An==> Cn2

(With this, it gets the TP)

3) An ? Cn =? (FP)

it returns a mosaic images with patches painted. Red regions are the TP. Blue regions are the FP, and the white squares are the original neuron patches in annot-file.