Add in patch the classes:  - SiftWeka  - instFeatures
I put the libraries which are necessary in the wiki.

The first class, SiftWeka.java calculates the SIFT features the batch of images (they are stored in different folders, if they are neurons, astrocytes... they are "the patches")
It stores these images with the SIFT features in other folder ("output folder").
Then, it uses weka to classified this patches, but at the moment THIS IS WRONG, because the instances are each point of each image (one instance should be all features of one patch). But I give that so, I think that later when the "bag of words" the change will be easy.

The second class, instFeatures.java, is a class for helping to build the instances for weka.